backend/requirements.txt
flask
flask-cors
opencv-python
mediapipe
yt-dlp
numpy>=1.26,<2
python-dotenv
